microservices/news_ingester/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
feedparser==6.0.10
google-cloud-pubsub==2.18.1
google-cloud-storage==2.10.0
pydantic==2.5.0